Refactor pipelines
