+++ src/tests/unit/pipelines/test_betting_pipeline.py
@@ -1,0 +1,13 @@
+from unittest import TestCase
+
+from kedro.pipeline import Pipeline
+
+from machine_learning.pipelines.betting_pipeline import create_betting_pipeline
+
+
+class TestBettingPipeline(TestCase):
+    def test_create_betting_pipeline(self):
+        pipeline = create_betting_pipeline("2000-01-01", "2010-12-31")
+
+        self.assertIsInstance(pipeline, Pipeline)
+        self.assertTrue(any(pipeline.nodes))
