+++ src/tests/unit/pipelines/test_player_pipeline.py
@@ -1,0 +1,13 @@
+from unittest import TestCase
+
+from kedro.pipeline import Pipeline
+
+from machine_learning.pipelines.player_pipeline import create_player_pipeline
+
+
+class TestPlayerPipeline(TestCase):
+    def test_create_player_pipeline(self):
+        pipeline = create_player_pipeline("2000-01-01", "2010-12-31")
+
+        self.assertIsInstance(pipeline, Pipeline)
+        self.assertTrue(any(pipeline.nodes))
